refactor(项目结构): 移除 SurgeryManager 类并调整项目命名空间
- 删除了 SurgeryManager.cs 文件
- 更新了命名空间：  - BodyPartUtils.cs 从 MyNamespace 改为 PrisonerManagementPanel.Utils
  - PawnSurgeryPolicyStorage.cs 从 PrisonerManagementPanel.Surgery 改为直接使用 namespace
- 移除了 SurgeryPolicy.cs 中未使用的 using 语句
- 在 RecipeFilterItem.cs 和 RecipeFilterUI.cs 中添加了对 PrisonerManagementPanel.Utils 的引用
