+++ PrisonerManagementPanel/Structure/RecipeFilterItem.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Verse;
+using PrisonerManagementPanel.Utils;
 
 namespace PrisonerManagementPanel.Structure;
