+++ PrisonerManagementPanel/Surgery/SurgeryPolicy.cs
@@ -1,5 +1,3 @@
-using System.Collections.Generic;
-using System.Linq;
 using RimWorld;
 using Verse;
